ランキング機能調整
