--- Assets/FishDetail_Ito.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-<<<<<<< HEAD:Assets/FishDetail_Ito.meta
-guid: 6670d2b63a472834e97745a96e680913
-folderAsset: yes
-DefaultImporter:
-=======
-guid: 7a3b1c540199a4495a90d2a96ba1bed0
-PrefabImporter:
->>>>>>> bca815199a87da1b14270a18a1a12b992fee53f7:Assets/One_OptionMenu.prefab.meta
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
