--- Assets/fish/itou_ver.4.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2d81ba31ede6bc346ad3bc44349a2606
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
