+++ Library/LastSceneManagerSetup.txt
@@ -1,5 +1,5 @@
 sceneSetups:
-- path: Assets/naichilab/unity-simple-ranking/Scenes/Ranking.unity
+- path: Assets/Scenes/ClearScene.unity
   isLoaded: 1
   isActive: 1
   isSubScene: 0
